Adding setup and requirements files.
