+++ requirements.txt
@@ -1,0 +1,24 @@
+appnope==0.1.2
+backcall==0.2.0
+certifi==2020.12.5
+decorator==4.4.2
+ipython==7.12.0
+ipython-genutils==0.2.0
+jedi==0.18.0
+mkl-fft==1.2.0
+mkl-random==1.1.1
+mkl-service==2.3.0
+numpy==1.19.2
+parso==0.8.1
+pexpect==4.8.0
+pickleshare==0.7.5
+pip==20.3.3
+prompt-toolkit==3.0.8
+ptyprocess==0.7.0
+pygame==2.0.1
+Pygments==2.7.3
+setuptools==51.0.0.post20201207
+six==1.15.0
+traitlets==4.3.3
+wcwidth==0.2.5
+wheel==0.36.2
